Create NavbarService
