+++ mitzu/webapp/service/navbar_service.py
@@ -1,0 +1,136 @@
+from __future__ import annotations
+
+import dash.development.base_component as bc
+
+import dash_bootstrap_components as dbc
+from dash import html
+from typing import List, Optional, cast
+import mitzu.webapp.pages.paths as P
+import flask
+import mitzu.webapp.dependencies as DEPS
+
+
+OFF_CANVAS_TOGGLER = "off-canvas-toggler"
+
+
+class NavbarService:
+    def __init__(self):
+        self._left_navbar_providers = []
+        self._right_navbar_provider = []
+
+        self._init_default_navbars()
+
+    def _init_default_navbars(self):
+        def off_canvas_toggle(id: str, **kwargs) -> Optional[bc.Component]:
+            off_canvas_toggler_visible = kwargs.get("off_canvas_toggler_visible", True)
+            return dbc.Button(
+                html.B(className="bi bi-list"),
+                color="primary",
+                size="sm",
+                className="me-3",
+                id={"type": OFF_CANVAS_TOGGLER, "index": id},
+                style={
+                    "display": "inline-block" if off_canvas_toggler_visible else "none"
+                },
+            )
+
+        def explore_button(id: str, **kwargs) -> Optional[bc.Component]:
+            create_explore_button = kwargs.get("create_explore_button", True)
+            storage = kwargs.get("storage", None)
+            project_name = kwargs.get("project_name", None)
+            if create_explore_button:
+                if storage is None:
+                    storage = cast(
+                        DEPS.Dependencies, flask.current_app.config.get(DEPS.CONFIG_KEY)
+                    ).storage
+
+                project_ids = storage.list_projects()
+                projects = [storage.get_project(p_id) for p_id in project_ids]
+                return dbc.DropdownMenu(
+                    children=[
+                        dbc.DropdownMenuItem(
+                            children=p.project_name,
+                            href=P.create_path(
+                                P.PROJECTS_EXPLORE_PATH, project_id=p.id
+                            ),
+                        )
+                        for p in projects
+                    ],
+                    size="sm",
+                    color="light",
+                    label="explore" if project_name is None else project_name,
+                    class_name="d-inline-block",
+                )
+            return None
+
+        self._left_navbar_providers = [
+            off_canvas_toggle,
+            explore_button,
+        ]
+
+        def signed_in_as(id: str, **kwargs) -> Optional[bc.Component]:
+            authorizer = cast(
+                DEPS.Dependencies, flask.current_app.config.get(DEPS.CONFIG_KEY)
+            ).authorizer
+
+            storage = cast(
+                DEPS.Dependencies, flask.current_app.config.get(DEPS.CONFIG_KEY)
+            ).storage
+
+            if authorizer is None:
+                return None
+
+            user_id = authorizer.get_current_user_id()
+            if user_id is None:
+                return None
+
+            email = None
+            if storage:
+                user = storage.get_user_by_id(user_id)
+                if user:
+                    email = user.email
+            if email is None:
+                email = user_id
+
+            return html.Div(
+                "Signed in as " + email,
+                style={"color": "white", "line-height": "2.4em", "font-weight": "bold"},
+            )
+
+        self._right_navbar_provider = [
+            signed_in_as,
+        ]
+
+    def get_navbar_component(
+        self, id: str, children: List[bc.Component] = [], **kwargs
+    ) -> dbc.Navbar:
+        navbar_comps = []
+
+        for provider in self._left_navbar_providers:
+            comp = provider(id, **kwargs)
+            if comp is not None:
+                navbar_comps.append(comp)
+
+        for comp in children:
+            navbar_comps.append(comp)
+
+        for provider in self._right_navbar_provider:
+            comp = provider(id, **kwargs)
+            if comp is not None:
+                navbar_comps.append(comp)
+
+        res = dbc.Navbar(
+            dbc.Container(
+                [
+                    dbc.Row(
+                        children=[dbc.Col(comp, width="auto") for comp in navbar_comps],
+                        className="g-2",
+                    ),
+                ],
+                fluid=True,
+            ),
+            class_name="mb-3",
+            color="dark",
+        )
+
+        return res
