+++ tests/unit/webapp/fixtures.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass, field
 import mitzu.webapp.dependencies as DEPS
 import mitzu.webapp.storage as S
 import mitzu.webapp.service.events_service as E
+import mitzu.webapp.service.navbar_service as NB
 import mitzu.model as M
 from mitzu.webapp.cache import MitzuCache
 import flask
@@ -50,6 +51,7 @@ def dependencies() -> DEPS.Dependencies:
         cache=cache,
         queue=queue,
         events_service=evt_service,
+        navbar_service=NB.NavbarService(),
         user_service=None,
     )
